@@ -13,7 +13,7 @@ trait EloquentUserTimezones
     {
         $value = parent::asDateTime($value);
 
-        if (!request()->is('nova*')) {
+        if (!request()?->is('nova*')) {
             $timezone = config('app.timezone_user');
 
             if (!empty($timezone)) {
@@ -29,7 +29,7 @@ trait EloquentUserTimezones
      */
     public function fromDateTime($value): ?string
     {
-        if ($value instanceof Carbon && !request()->is('nova*')) {
+        if ($value instanceof Carbon && !request()?->is('nova*')) {
             $timezone = config('app.timezone_user');
 
             if (!empty($timezone)) {
