@@ -6,7 +6,7 @@ use Illuminate\Auth\Middleware\Authenticate as IlluminateAuthenticate;
 
 class Authenticate extends IlluminateAuthenticate
 {
-    protected function authenticate($request, array $guards)
+    protected function authenticate($request, array $guards): void
     {
         parent::authenticate($request, $guards);
 
